# openapi.yaml
openapi: 3.0.0
info:
  title: Session API
  version: 1.0.0
  description: Returns all sessions.

servers:
  - url: https://copilot-api-1.onrender.com
    description: Public server

paths:
  /sessions:
    get:
      summary: Get all user sessions
      description: >
        Retrieves a list of all user sessions. Each session includes metadata such as name, start time, duration, summary, transcript, action items, and optional snapshots. 
        This endpoint is useful for answering questions about what meetings occurred, what was discussed, what action items were generated, and visual content shared during those meetings.
        The returned data can be filtered or summarized to support queries like:
        - "What did we talk about in the design meeting?"
        - "Show me my action items from recent sessions"
        - "What meetings happened last week?"
        - "Do we have any snapshots from the fundraiser planning meeting?"
      responses:
        '200':
          description: A list of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'

  /:
    get:
      summary: Health check
      description: Returns a simple message to confirm the API is running.
      operationId: getHealthCheck
      responses:
        '200':
          description: API is live
          content:
            text/plain:
              schema:
                type: string
  /sessions/summary:
    get:
      summary: Get summaries of all sessions
      description: Returns only the name, date, and summary of each session to answer high-level recap questions.
      responses:
        '200':
          description: A list of session summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    date:
                      type: string
                      format: date-time
                    summary:
                      type: string

  /sessions/recent:
    get:
      summary: Get sessions from the past N days
      description: >
        Returns sessions that started within the past N days. The number of days can be customized using the optional `days` query parameter. 
        Defaults to 7 if not provided.
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            minimum: 1
          description: Number of days to look back (e.g. 3 for past 3 days). Defaults to 7.
      responses:
        '200':
          description: A list of recent sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                  
  /sessions/actions:
    get:
      summary: Get all action items from all sessions
      description: Returns a flattened list of all action items across sessions, useful for questions like "what do I need to do this week?"
      responses:
        '200':
          description: A list of action items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    session:
                      type: string
                    actionText:
                      type: string
                    buttonAction:
                      type: string
                      nullable: true
                    description:
                      type: string
                      nullable: true
                    startTime:
                      type: string
                      format: date-time
                      nullable: true
                    endTime:
                      type: string
                      format: date-time
                      nullable: true
                    iconSrc:
                      type: string
                    location:
                      type: string
                      nullable: true
                    status:
                      type: integer
                    title:
                      type: string
                      nullable: true

  /sessions/by_attendee:
    get:
      summary: Get sessions by attendee
      description: Returns sessions where the specified person was present. Useful for queries like "What meetings did Alex attend?"
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
          description: Full name of the attendee (case-sensitive match).
      responses:
        '200':
          description: A list of sessions attended by the given person
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '400':
          description: Missing 'name' query parameter

  /sessions/by_date:
    get:
      summary: Get sessions within a date range
      description: >
        Returns sessions that occurred between the provided `from` and `to` dates (inclusive).
        Dates must be in ISO 8601 or YYYY-MM-DD format. Useful for answering questions like "What meetings happened in February?"
      parameters:
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date
          description: Start date (inclusive)
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date
          description: End date (inclusive)
      responses:
        '200':
          description: A list of sessions in the date range
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '400':
          description: Invalid or missing date format

  /sessions/topics:
    get:
      summary: Extract high-level topics from session summaries
      description: >
        Returns a list of topics mentioned across all session summaries. Capitalized keywords are extracted to represent main discussion areas.
        Useful for summarizing themes across meetings.
      responses:
        '200':
          description: A list of topic keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /sessions/{session_id}/{image_name}:
    get:
      summary: Serve a snapshot image from a session
      description: Returns an image file (PNG or JPG) from the specified session folder.
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
          description: The folder name for the session (e.g. "2025-03-10 Fundraiser...").
        - in: path
          name: image_name
          required: true
          schema:
            type: string
          description: The filename of the image to retrieve (e.g. "image_2025-03-10T16-03-23_UjFS.png").
      responses:
        '200':
          description: Snapshot image file
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found or invalid path



components:
  schemas:
    Session:
      type: object
      properties:
        sessionName:
          type: string
        date:
          type: string
          format: date
        attendees:
          type: array
          items:
            type: string
        summary:
          type: string
        transcript:
          type: string
        snapshots:
          type: array
          items:
            type: string